app/requirements_backend.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
asyncssh==2.19.0
atpublic==5.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.1
coverage==7.6.10
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
dpath==2.2.0
dulwich==0.22.7
dvc==3.59.0
dvc-data==3.16.8
dvc-gdrive==3.0.1
dvc-gs==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
funcy==2.0
gcsfs==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-run==0.10.14
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grandalf==0.8
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
gto==1.7.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
iniconfig==2.0.0
invoke==2.2.0
ipython==8.12.3
iterative-telemetry==0.0.9
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kagglehub==0.3.6
kiwisolver==1.4.8
kombu==5.4.2
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mpmath==1.3.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
narwhals==1.23.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
networkx==3.4.2
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.10.14
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
PyDrive2==1.21.3
pygit2==1.17.0
Pygments==2.19.1
pygtrie==2.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.9.1
safetensors==0.5.2
scikit-learn==1.3.0
scipy==1.15.1
scmrepo==3.3.9
semver==3.0.2
sentry-sdk==2.20.0
setproctitle==1.3.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqltrie==0.11.1
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
timm==1.0.14
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
voluptuous==0.15.2
wandb==0.19.4
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
win32_setctime==1.2.0
yarg==0.1.9
yarl==1.18.3
zc.lockfile==3.0.post1